requirements.txt
asgiref==3.2.3
certifi==2020.4.5.1
chardet==3.0.4
Django>=3.0.3,<4.0.0
django-appconf==1.0.4
django-compressor==2.4
django-filter==2.2.0
django-libsass==0.8
django-location-field==2.1.0
djangorestframework==3.11.0
geographiclib==1.50
geopy==1.21.0
idna==2.9
libsass==0.19.4
numpy==1.18.1
pandas==1.0.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
rcssmin==1.0.6
requests==2.23.0
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
